+++ camel/package.json
@@ -1,0 +1,13 @@
+{
+  "dependencies": {
+    "three": "^0.176.0"
+  },
+  "scripts": {
+    "dev": "vite",
+    "build": "vite build",
+    "preview": "vite preview"
+  },
+  "devDependencies": {
+    "vite": "^6.3.3"
+  }
+}
