--- camel/week2/README.md
@@ -1,11 +1,0 @@
-# practice-threejs
-
-재질 다운로드 사이트
-https://ambientcg.com/view?id=Rock058
-
-HDR 파일 다운로드 사이트
-https://polyhaven.com/a/autumn_field_puresky
-
-HDR vs EXR
-
-환경 맵은 비동기
