--- camel/week3/README.md
@@ -1,4 +1,0 @@
-# practice-threejs
-
-추천 3d ai 사이트
-https://vrin.co.kr/
